requirements.txt
torch      conda install pytorch==1.7.0 torchvision==0.8.0 torchaudio==0.7.0 cudatoolkit=11.0 -c pytorch
gensim
networkx
numpy
pyyaml
tensorboard
dgl     conda install -c dglteam dgl-cuda11.0
tqdm
tensorflow
matplotlib
scikit-learn


SHELL
conda install pytorch==1.7.0 torchvision==0.8.0 torchaudio==0.7.0 cudatoolkit=11.0 -c pytorch
conda install -c dglteam dgl-cuda11.0
conda install tensorflow gensim networkx pyyaml tensorboard tqdm matplotlib scikit-learn

nohup python -u main.py -G 1 -C config/long/DTW/config_50_50_long_qtree_full_embedding_32_dtw_Adam_0.001.yaml -W -P node2vec > outs/long/DTW/nor_Adam_0.001.out 2>&1 &
nohup python -u main.py -G 0 -C config/long/DTW/config_50_50_long_qtree_full_embedding_32_dtw_SGD_0.1.yaml -W -P node2vec > outs/long/DTW/nor_SGD_0.1.out 2>&1 &